Add support for depickling python datetime objects with dateutil.tz.tzutc objects as their timezone
This is tricky for a couple reasons. For protocol 0 and 1, Python didn’t use __setstate__ to unpickle objects implemented in C, it used functions from the copy_reg module. So, unlike pytz timezone objects which are implemented in python, the datetime.tzinfo standard library timezone class gets pickled using a “constructor” of copy_reg._reconstructor in protocol 0. Further complicating things is that the object that almost everyone uses, dateutil.tz.tzutc, is a subclass of this C class. This is pickled as a call to the _reconstructor function with three arguments: the subclass or “real” class we are trying to create, the base class which is implemented in C land, and the state that the eventual object we are trying to depickle should have. The reconstructor then allocates a new instance of the base class with __new__, but passes the sub class as the class to mark the object as. So the underlying C class does the allocation, but out comes an allocated instance of the fancy sub class. Yeeesh. Then, _reconstructor calls the base classes’ __init__ with the state to actually set the object back up again.

I opted to not really try to implement this logic cause IMO it’s crazy, and a deprecated protocol. Instead, I short cutted. I implemented a Pyrolite land version of the _reconstructor function which checks the “sub class” (given to it as a configured IObjectConstructor of some sort) to see if it knows how to reconstruct an object given the “base class” and “state”, which are from the unpickle stack. It uses reflection to check for a `reconstruct` function on the constructor. I then added an implementation of reconstruct to the TimeZoneConstructor so it can check for the very specific conditions that indicate we’re trying to unpickle a dateutil.tz.tzutc object, and just leverage the fact that we know it’s UTC and return a UTC TimeZone there. This is all, which means there is no support for unpickling other objects that are subclasses of C classes, like other things in dateutil, like dateutil.tz.tzlocal. This is something we could add in the future but requires making the code even more complex and confusing. I think it can be avoided for now since I am the first person to need this functionality, and I know I don’t need tzlocal support, since I feel as though most people stay away from those objects since they don’t support DST properly, and stick with pytz objects, which were easier to implement in the previous commit.

For protocol 2, __setstate__ is a thing, so I added a Tzinfo object which knows how to be depickled and have it’s state set. The DateTimeConstructor then knows how to be passed a Tzinfo object and ask it for its TimeZone so it can set that on the returned Calendar. Thing is, a dateutil.tz.tzutc object doesn’t actually have any state, womp womp. The timezone for it is simply implied by the name of the constructor as far as I can tell. This means that some object somewhere has to know that, so I did something somewhat dirty and make the tzinfo constructor understand when it needs to force itself to UTC. This means that the __setstate__ is called on the tzinfo object, but with no state, and has no effect, because the object previously learned what java-land TimeZone object it should represent when it was constructed, as that’s the only point in time where we knew the name of the constructor.

Yuck.