Improve pickling/unpickling performance
I found while using the library that it was generating a bunch of intermediate objects that could be avoided, and that these were contributing non-trivially to execution time.  I addressed many of them while also avoiding touching the public surface area, as I wasn't sure if it was maleable or not.